requirements.txt
# Core dependencies
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.25.0

# Audio processing
pydub>=0.25.1
librosa>=0.8.0
soundfile>=0.10.0
edge-tts>=6.1.3

# AI and ML
openai==0.27.8  # Using older version for httpx compatibility
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.19.0
demucs==4.0.1
spleeter==2.3.2

# Utilities
yt-dlp>=2023.0.0
beautifulsoup4>=4.9.0
